+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - Aug 4, 2025
+
+### Added
+
+- Add library import/export functionality - [(ISSUE)](https://github.com/nesaku/BiblioReads/issues/30)
+- Add a method to delete all library data
+- Add a type prop to select the toast type
+
+### Fixed
+
+- Fix the book cover and the placeholder cover both being visible at the same time
+
 ## [3.3.0] - Jun 11, 2025
 
 ### Added